Add files via upload
Since the beginning of this project, I've worked diligently to create and refine a website that embodies a minimalist and futuristic theme. Initially, I crafted a homepage featuring a top navigation bar, a hero section with an abstract image, and clear call-to-action buttons. As the project progressed, I made several stylistic adjustments to enhance the aesthetic appeal and functionality of the site, including implementing a darker color palette with bright accents for better visibility and modernizing the font styles.

I further enhanced the site's functionality by enabling smooth scrolling, allowing each navigation link to lead smoothly to its respective section. The contact section was equipped with a form that initially utilized PHP for email functionalities but was later integrated with Formspree to streamline the process.

Throughout the project, I addressed various layout and design issues, such as spacing, alignment, and the removal of text underlines to clean up the visual presentation. I also ensured that the website was responsive and accessible, adhering to best practices in web design. This involved troubleshooting issues like HTTP error 405 and making iterative improvements to enhance both the user interface and the overall user experience. These efforts have culminated in a cohesive and visually appealing website that aligns well with my futuristic branding goals.